backend/requirements.txt
alembic==1.7.4
anyio==3.3.4
asgi-lifespan==1.0.1
asgiref==3.4.1
async-generator==1.10
asyncpg==0.25.0
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
black==21.11b0
blessings==1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
coverage==6.2
decorator==5.1.0
fastapi==0.70.0
flake8==4.0.1
greenlet==1.1.2
h11==0.12.0
hexdump==3.3
httpcore==0.14.3
httptools==0.2.0
httpx==0.21.1
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.3.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.28.0
jedi==0.18.0
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
outcome==1.1.0
packaging==21.0
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.21
psycopg2==2.9.1
py==1.10.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==3.0.1
pyserial==3.5
pytest==6.2.5
pytest-asyncio==0.16.0
pytest-cov==3.0.0
python-dotenv==0.19.1
python-slugify==5.0.2
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.26
starlette==0.16.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
traitlets==5.1.0
trio==0.19.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.0
zipp==3.6.0
